hoario_empleado + check_in + check_out

+++ db/migrate/20151116222620_create_asistencias.rb
@@ -4,7 +4,7 @@ class CreateAsistencias < ActiveRecord::Migration
       t.boolean :entrada
       t.boolean :salida
       t.string :outputAsistencia
-      t.integer :horasTrabajadas
+      t.decimal :horasTrabajadas
 
       t.timestamps null: false
     end
